refactor: break up editor components
